decompositions: use float32 cumsum in masked_scatter to avoid INT64 stablehlo.reduce_window
ttnn does not support INT64 CumSumOp; the stablehlo.reduce_window with s64
type fails to legalize in StableHLO→TTIR. Switch mask_i to float32, compute
the cumsum in f32, then cast the result back to long for indexing.

+++ python_package/tt_torch/backend/decompositions.py
@@ -339,8 +339,9 @@ def masked_scatter(
     data_flat = data.reshape(-1)
     source_flat = source.reshape(-1)
 
-    mask_i = mask_f.long()
-    source_idx = torch.cumsum(mask_i, 0) - 1
+    # Use float32 cumsum: ttnn does not support INT64 cumsum (stablehlo.reduce_window s64 fails to legalize)
+    mask_i = mask_f.float()
+    source_idx = (torch.cumsum(mask_i, 0) - 1).long()
     source_idx = torch.clamp(source_idx, 0, source_flat.numel() - 1)
 
     gathered = source_flat[source_idx]
